+++ src/sane/core/display.hpp
@@ -20,7 +20,7 @@ static void key_callback(GLFWwindow* window, int key, int scancode, int action,
 namespace Sane {
   class Display {
   public:
-    Display(const char* Name, int Width, int Height);
+    Display(const char* Name, size_t Width, size_t Height);
     ~Display();
 
     void Update();
